requirements.txt
modal==0.56.4
diffusers>=0.14.0
transformers>=4.25.1
accelerate>=0.16.0
torch>=2.0.0
Pillow>=9.3.0
fastapi>=0.95.0
python-dotenv>=1.0.0 